more refactor

+++ src/main/java/edu/utn/mail/controller/MessageController.java
@@ -35,7 +35,7 @@ public class MessageController {
         return messageService.newMessage(message);
     }
 
-    public Message getMessage(Integer messageId) {
+    public Message getMessage(Long messageId) {
         return messageService.getMessageById(messageId);
     }
 }
